fix: 修复 ClickHouse DateTime 类型序列化问题和启动脚本
- 修复 stock_realtime_quotes 表 DateTime 序列化错误
  * 将 timestamp 字段从 DateTime 改为 UInt64
  * Rust 代码中使用 u64 类型存储 Unix 时间戳
  * 查询时使用 toDateTime(timestamp, 'Asia/Shanghai') 转换
  * 添加缺失的 market 字段

- 修复 start-all.sh 启动脚本
  * 明确指定运行 data-collector 二进制文件
  * 解决 cargo run 无法确定二进制文件的问题

- 新增文档
  * docs/TROUBLESHOOTING.md - 完整的故障排查指南
  * 更新 DEPLOYMENT_TEST_REPORT.md - 真实数据测试报告
  * 更新 README.md - 添加新文档链接

测试验证：
- ✅ 环境清空和重新部署成功
- ✅ 所有服务正常启动并运行
- ✅ 数据采集和存储功能正常
- ✅ 无严重 bugs

+++ services/data-collector/src/adapters/secondary/clickhouse_repository.rs
@@ -22,7 +22,7 @@ impl ClickHouseQuoteRepository {
 
     /// Convert legacy StockQuote to domain StockQuote
     fn legacy_to_domain(&self, legacy: &LegacyStockQuote) -> Result<StockQuote, String> {
-        let timestamp = DateTime::from_timestamp(legacy.timestamp, 0)
+        let timestamp = DateTime::from_timestamp(legacy.timestamp as i64, 0)
             .unwrap_or_else(|| Utc::now());
         let code = StockCode::new(legacy.code.clone())?;
         let price = Price::new(legacy.price)?;
@@ -49,7 +49,7 @@ impl ClickHouseQuoteRepository {
     /// Convert domain StockQuote to legacy StockQuote for ClickHouse
     fn domain_to_legacy(&self, domain: &StockQuote) -> LegacyStockQuote {
         LegacyStockQuote {
-            timestamp: domain.timestamp.timestamp(),
+            timestamp: domain.timestamp.timestamp() as u64,
             code: domain.code.as_str().to_string(),
             name: domain.name.clone(),
             price: domain.price.value(),
